Web Update | 11.10.25 | 11:11 PM
Changes Made: 
Separated CSS and JS code within HTML files into their own files to improve the website's organization, performance, and maintenance. 
Changed the Hamburger Menu from an icon to bars with an animation that turns into an "X" when the user wants to exit.
Changed the Hamburger Menu dropdown and hover colors.
Changed the Website's colors to match the wireframe and design guide.
Added Icons and Badge Icon to Achievements.
Corrected the usage of "Assignments" to "Achievements"..
Tried to match achievements progress bar to match wireframe progress bar as close as I could. 
Made header sticky, so when user scrolls it stays with them for easy access to navbar menu.