Web Update | 11.21.2025 | 1:23 AM
Had to rewrite some files so adapting the wireframe to the website would be easier.
- home.html ---> homepage.html
- photos.html ---> trails.html
- style.css ---> gtgstyle.css
- gtgweb.js ---> gtgjs.js

Added classes and id attributes to tags so not every <section> had the same style and was changeable for a different design. 
Renamed some names to make editing CSS easier to understand and use.
Added comments for easier read.
Added slideshow feature on homepage.html (not completely finished).